Update README.md
This project is a collection of Solidity-based smart contracts designed for various use cases. It utilizes Hardhat, a popular Ethereum development environment, for compiling, testing, and deploying smart contracts. With built-in testing and gas reporting features, this project helps ensure that your smart contracts are both functional and optimized for efficiency.

The purpose of this repository is to provide developers with a robust foundation for building decentralized applications (DApps) and to serve as an example for working with Solidity, Hardhat, and gas optimization.

Key features of this repository include:

Smart Contract Development: Includes various Solidity contracts for use and study.
Testing: Hardhat and Chai are used to write automated tests for ensuring contract correctness.
Gas Optimization: Provides gas usage reporting to help optimize contract execution costs.
This repository is ideal for developers who are either starting with Ethereum development or looking to improve their smart contract testing and optimization practices.